requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
python-dotenv>=1.0.0
pydantic>=2.0.0
streamlit>=1.28.0
ast-comments>=1.1.1
bandit>=1.7.5
pylint>=3.0.0
black>=23.0.0
radon>=6.0.0

